Controller & Directive Approach Changes
SourceDetail State Changes will no longer cause the controller to
reload (unless the course itself changes) . The controller will catch
changes to the detail and apply a change to $scope.detailCard. This
change will be caught by the detail directive and affect a reload of
the detail content. This means the rest of the view (list/stream view)
remains unaffected by the state change (except that it can of course
read the latest scope changes and update accordingly (for selected
classes, etc)